tambah debug

+++ njajal-3/run-ns.sh
@@ -1,2 +1,9 @@
+if [ -f trace.log ]
+then
 rm trace.log
+fi
+if [ -f debug.log ]
+then
+rm debug.log
+fi
 ns njajal-aodv.tcl
